Properly implement purging
When many Ritz values have converged, but suddenly new Ritz values
closer to the target are detected (which happens often with repeated
eigenvalues close to the target), it may happen that num_locked +
num_new > nev.

In that case we have to purge old vectors. Those old vectors may be
anywhere in the range `1:active-1`, so we have to ensure that the change
of basis of `V` is from `V[:, purge:k]` instead of `V[:, active:k]`.